+++ react-components_1_week/src/components/formCarts/FormCarts.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { carts } from 'components/pages/FormsData';
+import CreatFormCart from './CreatFormCart';
+
+export default class FormCarts extends React.Component {
+  render() {
+    if (carts.length !== 0) {
+      return (
+        <div className="form-carts">
+          {carts.map((item, key) => {
+            return <CreatFormCart card={item} key={key} />;
+          })}
+        </div>
+      );
+    }
+  }
+}
